commit orderDetailResponse

+++ BackEnd/shopapp/src/main/java/com/project/shopapp/dtos/OrderDetailDTO.java
@@ -15,13 +15,13 @@ public class OrderDetailDTO {
   @Min(value = 1, message = "Product's ID must be > 0")
   private Long productId;
   @Min(value = 0, message = "Price must be >= 0")
-  private Long price;
+  private Float price;
   @JsonProperty("number_of_products")
   @Min(value = 1, message = "Number of products must be >= 1")
   private int numberOfProducts;
   @JsonProperty("total_money")
   @Min(value = 0, message = "Total money must be > 0")
-  private int totalMoney;
+  private float totalMoney;
   public String color;
 
 }
